+++ src/molflux/metrics/regression/root_mean_squared_error.py
@@ -1,7 +1,6 @@
-"""Mean squared error regression loss."""
+"""Root mean squared error regression loss."""
 
 import logging
-import warnings
 from importlib.metadata import version
 from typing import Any, List, Optional
 
@@ -24,8 +23,8 @@ from molflux.metrics.typing import ArrayLike, MetricResult
 logger = logging.getLogger(__name__)
 
 _DESCRIPTION = """
-The mean_squared_error function computes mean square error, a risk metric
-corresponding to the expected value of the squared (quadratic) error or loss.
+The root_mean_squared_error function computes root mean square error, a risk metric
+corresponding to the expected value of the (quadratic) error or loss.
 """
 
 _KWARGS_DESCRIPTION = """
@@ -40,35 +39,23 @@ Args:
             Returns a full set of errors in case of multioutput input.
         'uniform_average' :
             Errors of all outputs are averaged with uniform weight.
-    root (optional): If ``True`` returns RMSE value, if ``False`` returns MSE
-        value. Defaults to ``False``.
-
-        .. deprecated:: 0.23.0
-           `root` is deprecated in 0.23.0 and will be removed in future releases.
-           Use the `root_mean_squared_error` metric instead to calculate the root mean squared error.
 
 Returns:
-    mean_squared_error: A non-negative floating point value (the best value
+    root_mean_squared_error: A non-negative floating point value (the best value
     is 0.0), or an array of floating point values, one for each individual target.
 
 Examples:
     >>> from molflux.metrics import load_metric
-    >>> metric = load_metric("mean_squared_error")
+    >>> metric = load_metric("root_mean_squared_error")
     >>> predictions = [2.5, 0.0, 2, 8]
     >>> references = [3, -0.5, 2, 7]
     >>> metric.compute(predictions=predictions, references=references)
-    {'mean_squared_error': 0.375}
-    >>> references = [2.5, 0.0, 2, 8]
-    >>> predictions = [3, -0.5, 2, 7]
-    >>> metric = load_metric("mean_squared_error", config_name="multioutput")
+    {'root_mean_squared_error': 0.612...}
+    >>> metric = load_metric("root_mean_squared_error", config_name="multioutput")
     >>> predictions = [[0, 2],[-1, 2],[8, -5]]
     >>> references = [[0.5, 1],[-1, 1],[7, -6]]
     >>> metric.compute(predictions=predictions, references=references)
-    {'mean_squared_error': 0.708...}
-    >>> metric.compute(predictions=predictions, references=references, multioutput='raw_values')
-    {'mean_squared_error': array([0.416..., 1.        ])}
-    >>> metric.compute(predictions=predictions, references=references, multioutput=[0.3, 0.7])
-    {'mean_squared_error': 0.825...}
+    {'root_mean_squared_error': 0.822...}
 """
 
 _CITATION = """\
@@ -87,7 +74,7 @@ _CITATION = """\
 
 
 @evaluate.utils.file_utils.add_start_docstrings(_DESCRIPTION, _KWARGS_DESCRIPTION)
-class MeanSquaredError(HFMetric):
+class RootMeanSquaredError(HFMetric):
     def _info(self) -> evaluate.MetricInfo:
         return evaluate.MetricInfo(
             description=_DESCRIPTION,
@@ -105,7 +92,7 @@ class MeanSquaredError(HFMetric):
                 },
             ),
             reference_urls=[
-                "https://scikit-learn.org/stable/modules/generated/sklearn.metrics.mean_squared_error.html",
+                "https://scikit-learn.org/stable/modules/generated/sklearn.metrics.root_mean_squared_error.html",
             ],
         )
 
@@ -116,49 +103,22 @@ class MeanSquaredError(HFMetric):
         references: ArrayLike,
         sample_weight: Optional[List[float]] = None,
         multioutput: str = "uniform_average",
-        root: bool = "deprecated",  # type: ignore[assignment]
         **kwargs: Any,
     ) -> MetricResult:
-        root_kwarg_set_by_user = root != "deprecated"
-        if root_kwarg_set_by_user:
-            warnings.warn(
-                (
-                    "'root' is deprecated in version 0.24.0 and "
-                    "will be removed in future versions. To calculate the "
-                    "root mean squared error, use the metric "
-                    "'root_mean_squared_error'."
-                ),
-                FutureWarning,
-                stacklevel=1,
-            )
-        else:  # swap in with actual default value
-            root = False
-
         # TODO(avianello): cleanup once we can drop python 3.8 (RMSE will always be available)
         if ROOT_MEAN_SQUARED_ERROR_AVAILABLE:  # py3.9+
-            if root:
-                score = root_mean_squared_error(
-                    y_true=references,
-                    y_pred=predictions,
-                    sample_weight=sample_weight,
-                    multioutput=multioutput,
-                )
-
-            else:
-                score = mean_squared_error(
-                    y_true=references,
-                    y_pred=predictions,
-                    sample_weight=sample_weight,
-                    multioutput=multioutput,
-                )
-
+            score = root_mean_squared_error(
+                y_true=references,
+                y_pred=predictions,
+                sample_weight=sample_weight,
+                multioutput=multioutput,
+            )
         else:  # py3.8
             score = mean_squared_error(
                 y_true=references,
                 y_pred=predictions,
                 sample_weight=sample_weight,
                 multioutput=multioutput,
-                squared=not root,
+                squared=False,
             )
-
         return {self.tag: score}