+++ src/molflux/metrics/suites/discovery.py
@@ -7,10 +7,6 @@ _CWD: Path = Path(__file__).parent.resolve()
 
 def import_suites() -> None:
     """Load all available suites in the catalogue."""
-    paths = (
-        path
-        for path in _CWD.iterdir()
-        if path.suffix == ".yml" or path.suffix == ".yaml"
-    )
+    paths = (path for path in _CWD.iterdir() if path.suffix in {".yml", ".yaml"})
     for path in paths:
         register_suite(suite_path=path)